util: Move socketset from net resource to util
Prepare the socketset api to be used outside of the scope the net
resource.
